ST-294:Moved Journey cache logic to a separate trait
