+++ app/uk/gov/hmrc/tai/cache/estimatedPay/UpdatedEstimatedPayJourneyCache.scala
@@ -1,0 +1,50 @@
+/*
+ * Copyright 2018 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.tai.cache.estimatedPay
+
+import uk.gov.hmrc.http.HeaderCarrier
+import uk.gov.hmrc.tai.service.JourneyCacheService
+import uk.gov.hmrc.tai.util.JourneyCacheConstants
+
+import scala.concurrent.{ExecutionContext, Future}
+
+trait UpdatedEstimatedPayJourneyCache extends JourneyCacheConstants{
+
+  def journeyCache(key: String, cacheMap: Map[String, String])
+                  (implicit hc: HeaderCarrier, ec: ExecutionContext, journeyCacheService: JourneyCacheService): Future[Map[String, String]] = {
+
+    def updateCache(cacheToUpdate: Map[String,String],keysToEmpty: List[String])(implicit hc: HeaderCarrier): Future[Map[String, String]] = {
+      if (cacheMap(key) == "Yes") {
+        journeyCacheService.cache(cacheMap)
+      } else {
+        for{
+          current <- journeyCacheService.currentCache
+          updatedCacheMap = current.filterKeys( key => !keysToEmpty.contains(key)) ++ cacheToUpdate
+          _ <- journeyCacheService.flush()
+          updatedCache <- journeyCacheService.cache(updatedCacheMap)
+        }yield updatedCache
+      }
+    }
+
+    key match {
+      case UpdateIncome_PayslipDeductionsKey => updateCache(Map(key -> cacheMap(key)), List(UpdateIncome_TaxablePayKey))
+      case UpdateIncome_BonusPaymentsKey => updateCache(Map(key -> cacheMap(key)),
+        List(UpdateIncome_BonusPaymentsThisYearKey, UpdateIncome_BonusOvertimeAmountKey))
+      case _ => journeyCacheService.cache(cacheMap)
+    }
+  }
+}
